% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_Config.R
\name{.OpenConfigFile}
\alias{.OpenConfigFile}
\title{.OpenConfigFile
Internal. save and restore config info}
\usage{
.OpenConfigFile(config)
}
\arguments{
\item{config}{config list structure with software$softName...}
}
\description{
.OpenConfigFile
Internal. save and restore config info
}
