% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_installers.R
\name{InstallMiXCR}
\alias{InstallMiXCR}
\title{InstallMiXCR}
\usage{
\dontrun{
InstallMiXCR("myfavoriteplace")
}
}
\arguments{
\item{where}{(character) the path to the directory where I want to install MiXCR}
}
\description{
The software MiXCR will be instaled at ./home/where/Software/mixcr-3.0.13/
if the directory "where" do not exists, the it will be created as well as directory "Software/mixcr-3.0.13/" below "where"
so it will look like

./home/

......where/

...........Software/mixcr-3.0.13/
}
