% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCR}
\alias{RunMiXCR}
\title{RunMiXCR}
\usage{
RunMiXCR(
  sbj,
  species = c("hsa", "mmu"),
  assemblePartial = FALSE,
  extendedAlignments = FALSE,
  nThreads
)
}
\arguments{
\item{sbj}{full path to sample fastq R1 and R2. It should be named as xxxx_R1.fastq and xxxx_R2.fastq}

\item{species}{any of "hsa","mmu" (human - mouse)}

\item{assemblePartial}{(logical, dafault FALSE) if parial read should be assembled}

\item{extendedAlignments}{(logical, dafault FALSE) if TCR extension alignment should be performed}

\item{nThreads}{(integer) number of cores}
}
\value{
a list with clones, vmin, vdetails.clones, TRA.clones,TRB.clones,TRG.clones,TRD.clones data frames
it also save an Excel file as full.path.sbj.alignments.vdjca.xlsx
}
\description{
RunMiXCR
}
