% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRreport}
\alias{RunMiXCRreport}
\title{RunMiXCRreport}
\usage{
RunMiXCRreport(
  clonsFile,
  type = c("full", "min"),
  otherFields,
  fileType = c("Excel", "tab")
)
}
\arguments{
\item{clonsFile}{the full path to the clone file}

\item{type}{string "full" (default) or "min"}

\item{otherFields}{other fileds (not implemented yet)}
}
\description{
RunMiXCRreport
}
