% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRseqAlignment}
\alias{RunMiXCRseqAlignment}
\title{RunMiXCR.RNAseqAlignment}
\usage{
RunMiXCRseqAlignment(
  sbj,
  species = c("hsa", "mmu", "rat"),
  extendedAlignments = FALSE,
  nThreads = 4L
)
}
\arguments{
\item{sbj}{full path to sample fastq R1 and R2. It should be named as xxxx_R1.fastq and xxxx_R2.fastq}

\item{species}{any of "hsa","mmu", "rat" (human - mouse, rat)
#'@param extendedAlignments (logical, default TRUE) if TCR extension alignment should be performed}

\item{nThreads}{(integer) number of cores (default 4, since it is suggested as the optimum value)}
}
\value{
the output file full path name.
}
\description{
RunMiXCR.RNAseqAlignment
}
