% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRexportExcel}
\alias{RunMiXCRexportExcel}
\title{RunMiXCRexportExcel}
\usage{
RunMiXCRexportExcel(clonsFile)
}
\arguments{
\item{clonsFile}{(string) the clone file. No IG clones}
}
\description{
RunMiXCRexportExcel
}
