% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRseq}
\alias{RunMiXCRseq}
\title{RunMiXCRseq
It runs the MiXCR package for RNAseq experiments}
\usage{
RunMiXCRseq(sbj, species = c("hsa", "mmu"), extendedAlignments = TRUE)
}
\arguments{
\item{sbj}{full path fastq(.gz) R1 sample}

\item{species}{character any of "hsa","mmu" (default hsa)}
}
\value{
if success, the it builds two files sbj.alignments.vdjca and sbj.clones.clns and return a vector with both full path file names
if fail, returns NA
}
\description{
RunMiXCRseq
It runs the MiXCR package for RNAseq experiments
}
