% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRseqAlignment}
\alias{RunMiXCRseqAlignment}
\title{RNAseqAlignment}
\usage{
RunMiXCRseqAlignment(sbj, species = c("hsa","mmu","rat"), extendedAlignments=FALSE, nThreads = 4L)
}
\arguments{
\item{sbj}{full path to the subject.R1.fastq sample file. The paired read should be named subject.R2.fastq}

\item{species}{any of "hsa","mmu", "rat" (human - mouse, rat)}

\item{extendedAlignments}{(logical, default TRUE) if TCR extension alignment should be performed (in case of incomplete TCR alignments)}

\item{nThreads}{(integer) number of cores (default 4, since it is suggested as the optimum value)}
}
\value{
It will generate file named "subject.R1.fastq.alignments.vdjca"
the full path to this file is returned to be used as input to \code{\link{RunMiXCRseqClones}}
}
\description{
This function aligns raw sequencing reads to reference V, D, J and C genes of T- and B- cell receptors.
}
\examples{
/dontrun{
      subj.file <- "/home/.../subject.R1.fastq"
      out.vdja.file <- RunMiXCR(sbj = subj.file)
      print(out.vdja.file)
}


}
