% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMiXCR_runners.R
\name{RunMiXCRseqClones}
\alias{RunMiXCRseqClones}
\title{RunMiXCR.RNAseqClones}
\usage{
RunMiXCRseqClones(alignmentFile, nThreads = 4L)
}
\arguments{
\item{alignmentFile}{full path to sample fastq R1 and R2. It should be named as xxxx_R1.fastq and xxxx_R2.fastq}

\item{nThreads}{(integer) number of cores (4 default as it is suggested as the optimal value)}
}
\value{
the full path name of the clones file
}
\description{
RunMiXCR.RNAseqClones
}
